+++ package.json
@@ -14,6 +14,7 @@
     "@types/uuid": "^8.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
     "tailwindcss": "^3.2.1",
     "typescript": "*",
