+++ src/components/map/SelectedDot.tsx
